Return 302 status from 404 page
The 404 page is being used as a catch-all page to catch all incoming requests to go.param.me/xxx. It doesn't make sense to respond with a 404 Not Found status, because it is redirecting the client to a different page.

Currently, http-equiv status is an HTML extension specification and only defines the 410 Gone value, but allows any status code. See https://indieweb.org/meta_http-equiv_status